Cleaned headers
Changed the headers and code to fit the more typical format of the Interface/Implementation stuff.
